Refactor: rename component into JS camelCase
