+++ onboarding/src/Components/AddUserTable/AddUserTable.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import Users from "../hooks/Users";
-import AddUserTableRow from "./Add_User_Table_Row";
+import AddUserTableRow from "./AddUserTableRow";
 //import UserListSearch from "../UserListSearch";
 import { assignEmployeeToPackage } from "../hooks/EmployeeForms";
 import UserListSearch from "../UserListSearch";
